Simplify documentation for search paths
##### SUMMARY

I found this information incredibly hard to read and understand. I tried my best to make it easier to read but I don't actually know how this works - if I knew how it worked, I wouldn't have searched for this documentation in the first place. So really this is just my best attempt at interpreting and translating the information that was already here.

If you know better, please do point out any mistakes I made. I'll be happy to rewrite it.


##### ISSUE TYPE
- Docs Pull Request

+label: docsite_pr